COMP: cast size_t to int before return to avoid compiler warning
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=17499
git-svn-id: http://svn.slicer.org/Slicer4/trunk@17499 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/vtkMRMLSelectionNode.h
@@ -144,7 +144,7 @@ class VTK_MRML_EXPORT vtkMRMLSelectionNode : public vtkMRMLNode
   std::string GetAnnotationResourceByID(std::string id);
 //ETX
   /// Get the number of ids in the list
-  int GetNumberOfAnnotationIDsInList() { return this->AnnotationIDList.size(); };
+  int GetNumberOfAnnotationIDsInList() { return (int)(this->AnnotationIDList.size()); };
 protected:
   vtkMRMLSelectionNode();
   ~vtkMRMLSelectionNode();
